backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# AI & LangChain
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.10
langgraph==0.0.20
google-generativeai==0.3.2

# Vector Database
pinecone-client==3.0.2

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
pypdf==4.0.1

# Authentication (Optional)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
httpx==0.26.0
aiofiles==23.2.1
boto3==1.34.34  # For S3 storage
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Monitoring & Logging
python-json-logger==2.0.7
